feat(projen.project.typescript): refactor and move typescript project to subpackage.
