+++ packages/projen/project/typescript/src/index.ts
@@ -1,0 +1,2 @@
+export * from './typescript'
+export * from './typescript-project-options'
